backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg[binary]==3.1.12
psycopg-pool==3.1.7
google-generativeai>=0.8.0
python-multipart==0.0.6
werkzeug==2.3.7